+++ packages/interweave-emoji-picker/tests/mocks.js
@@ -3,5 +3,8 @@ import { CONTEXT_CLASSNAMES, CONTEXT_MESSAGES } from '../src/constants';
 // eslint-disable-next-line
 export const PICKER_CONTEXT = {
   classNames: CONTEXT_CLASSNAMES,
-  messages: CONTEXT_MESSAGES,
+  messages: {
+    ...CONTEXT_MESSAGES,
+    noPreview: 'No emoji to display',
+  },
 };